Added to Canvas directory, fixed list items to generate in loop, alphabetized imports
